samples: wifi: radio_test: Move shell into nRF70 and nRF71 drivers
Move radio test shell commands out of the radio_test sample and into
the nRF70 and nRF71 Wi-Fi drivers. Register the nRF71 command set
behind CONFIG_NRF71_RADIO_TEST_SHELL and slim the radio_test sample
to application entry points only. Remove the shared shell and FICR
sources from samples/wifi/radio_test/common.

Bring in the matching nRF70 radio test and FICR shell support to
sdk-zephyr.
