+++ samples/wifi/radio_test/single_domain/CMakeLists.txt
@@ -9,29 +9,12 @@ cmake_minimum_required(VERSION 3.20.0)
 find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
 project(wifi_ble_radio_test)
 
-set(common_rt_dir ${ZEPHYR_NRF_MODULE_DIR}/samples/wifi/radio_test/common)
 set(perip_rt_dir ${ZEPHYR_NRF_MODULE_DIR}/samples/peripheral/radio_test)
-
-if(CONFIG_NRF71_RADIO_TEST)
-  target_link_libraries(app PRIVATE nrf71_wifi_radio_test_headers)
-endif()
-
-target_include_directories(app PRIVATE ${common_rt_dir}/inc)
 
 file(GLOB perip_rt_sources ${perip_rt_dir}/src/*.c)
 
 # NORDIC SDK APP START
 target_sources(app PRIVATE
   ${perip_rt_sources}
-  ${common_rt_dir}/src/nrf_wifi_radio_test_shell.c
 )
-
-if(NOT CONFIG_NRF71_RADIO_TEST)
-  target_sources(app PRIVATE
-    ${common_rt_dir}/src/nrf_wifi_radio_ficr_shell.c
-    ${common_rt_dir}/src/ficr_prog.c
-  )
-
-  target_link_libraries(app PRIVATE nrf-wifi-osal)
-endif()
 # NORDIC SDK APP END
